// src/app/shared/model/flight.ts
export class Flight {
    flightNo: number;
    carrierName: string;
    flightModel: string;
    seatCapacity: number;
    action: boolean;
    userId: number;
    price: number;
    id: number;
    userIdUpdate: number;
    createdDate: Date;
    origin: string;
    desitnation: string;
    activeStatus: string;
}